v2.1.2
Updating version numbers.  Updating Environments to be centered and visible in tests.  Updated Backport and other UI functions that should be version-restricted.  Tested Apple Watch version (missed in 2.1 testing).  Fixed issue with storage Ints being too small (so specified Int64).  Fixed updates for macCatalyst.

+++ Sources/DeviceKitBridge.swift
@@ -451,13 +451,13 @@ extension Device {
 extension Device {
     /// The volume’s total capacity in bytes.
     @available(*, deprecated, renamed: "Device.current.volumeTotalCapacity")
-    public static var volumeTotalCapacity: Int? {
+    public static var volumeTotalCapacity: Int64? {
         Device.current.volumeTotalCapacity
     }
     
     /// The volume’s available capacity in bytes.
     @available(*, deprecated, renamed: "Device.current.volumeAvailableCapacity")
-    public static var volumeAvailableCapacity: Int? {
+    public static var volumeAvailableCapacity: Int64? {
         Device.current.volumeAvailableCapacity
     }
